Phase 2.4: Add SIMD to Feedforward & Fix Float→Double
✅ COMPLETED:
- Added SIMD optimizations to cllm_feedforward.c
  * AVX2 vectorized ReLU activation (4 doubles at a time)
  * AVX2 vectorized GELU activation with fast approximation
  * AVX2 vectorized matrix-vector multiplication
  * Implemented cllm_feedforward_forward() with full SIMD
  * Implemented cllm_feedforward_batch() for batch processing
  * 2-4x speedup potential over scalar implementation

- Fixed float→double consistency in CLLMToken structure
  * Changed lattice_coords from float[3] to double[3]
  * Changed angle from float to double
  * Changed radius from float to double
  * Changed frequency from float to double
  * Ensures consistent precision throughout codebase

🎯 SIMD OPTIMIZATIONS NOW COMPLETE:
- ✅ Attention (AVX2 matrix multiplication, softmax)
- ✅ Layer normalization (AVX2 mean, variance, normalization)
- ✅ Feedforward (AVX2 ReLU, GELU, matrix-vector)
- ✅ All using double precision consistently

📊 PERFORMANCE EXPECTATIONS:
- ReLU: 4x speedup (simple max operation)
- GELU: 2-3x speedup (approximation + vectorization)
- Matrix-vector: 2-4x speedup (FMA instructions)
- Overall feedforward: 2-4x speedup

🎯 NEXT:
- Run performance benchmarks
- Complete memory layout optimization
- Continue with master plan

Status: SIMD optimization complete, clean build maintained

+++ include/cllm.h
@@ -157,11 +157,11 @@ typedef struct {
 typedef struct {
     uint32_t token_id;
     uint64_t prime_encoding;    // Prime number encoding
-    float lattice_coords[3];    // 3D lattice coordinates
-    float angle;                // Angular position
-    float radius;               // Radial distance
+    double lattice_coords[3];   // 3D lattice coordinates
+    double angle;               // Angular position
+    double radius;              // Radial distance
     char token_str[64];
-    float frequency;
+    double frequency;
     uint32_t symmetry_group;    // 0 to 11 (12-fold)
     uint8_t reserved[20];
 } CLLMToken;
